Modify enum to components
